Made the game work!
The buttons toggle a random set of other buttons, making the aim of the game to turn all the buttons on. The central button randomizes each time to make the game always possible. It isn't preferable to the other buttons however, as there is about a 1/32000 chance that it will turn all other buttons on (or leave them on).